Add Fedora 44 GPU-PV support and fix 16 critical/high-severity bugs
## Fedora 44 Support
- New Fedora-44-Official.sh deployment script (644 lines)
  - dnf package management, CachyOS kernel detection
  - Secure Boot MOK key generation + module signing
  - SELinux context labels for /usr/lib/wsl/lib/
  - dracut initramfs, Mesa from base repos (no PPA needed)
  - Uninstall support and verification step
- Updated index.json with Fedora script entry

## Critical Bug Fixes (8)
- RemoveGpuPartitionAsync: now properly awaits Run2 (always returned true)
- VmCreateService: integer overflow in memory calc at 2048+ MB
- SshService: password shell injection via unescaped single quotes
- NvidiaReg: blind Replace("DriverStore") corrupted registry values
- NvidiaReg: return value was silently discarded
- WmiTools.WaitForJob: infinite loop replaced with 5-min timeout
- SystemSwitcher: ACCESS_DENIED (code 5) no longer treated as success
- HcsManager: job state loop fixed (was looping on Completed state)

## High-Severity Fixes (8)
- Utils.Run(): now logs PowerShell errors via Debug.WriteLine
- VmGPUService.ExecuteCommand: Process resource leak (using + pipe drain)
- UsbVmbusService: removed dangerous RealTime process priority
- VM power-off polling: 2-minute timeout added (was infinite loop)
- Robocopy exit code now checked (>=8 = error)
- LastOrDefault().Id NullReferenceException prevented
- GPU removal error snackbar: fixed wrong resource key
- GPU rollback: wrapped in try/catch to prevent masking errors

+++ src/Services/VmCreateService.cs
@@ -75,7 +75,7 @@ namespace ExHyperV.Services
                         switchParam = $"-SwitchName '{p.SwitchName}'";
                     }
 
-                    long memoryBytes = p.MemoryMb * 1024 * 1024;
+                    long memoryBytes = (long)p.MemoryMb * 1024L * 1024L;
                     string diskParam = p.DiskMode switch
                     {
                         0 => $"-NewVHDPath '{p.VhdPath}' -NewVHDSizeBytes {p.DiskSizeGb}GB",
